catalog-backend/requirements.txt
Flask==1.0.2
itsdangerous==1.1.0
requests==2.18.4
flask_cors==3.0.7
flask_httpauth==3.2.4
httplib2==0.12.0
passlib==1.7.1
redis==3.1.0
sqlalchemy
psycopg2-binary
oauth2client
fake-data
packaging
flask-sqlalchemy
bleach